refactor(package): 패키지를 domain과 ui로 구분한다
