+++ src/main/java/ladder/domain/User.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.domain;
 
 public class User {
     private final String userName;
