+++ src/test/java/ladder/UserTest.java
@@ -1,5 +1,6 @@
 package ladder;
 
+import ladder.domain.User;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
